move autooai into outer directory
